+++ src/main/java/edu/cornell/kfs/rass/batch/service/RassSortService.java
@@ -1,0 +1,10 @@
+package edu.cornell.kfs.rass.batch.service;
+
+import java.util.List;
+
+import edu.cornell.kfs.rass.batch.xml.RassXmlAgencyEntry;
+
+public interface RassSortService {
+    List<RassXmlAgencyEntry> sortRassXmlAgencyEntriesForUpdate(List<RassXmlAgencyEntry> agencies);
+
+}
